+++ plugins/woocommerce-admin/packages/components/src/advanced-filters/index.js
@@ -117,7 +117,7 @@ class AdvancedFilters extends Component {
 		onAdvancedFilterAction( 'match', { match } );
 	}
 
-	onFilterChange( index, property, value, shouldResetValue = false ) {
+	onFilterChange( index, { property, value, shouldResetValue = false } ) {
 		const newActiveFilters = [ ...this.state.activeFilters ];
 		newActiveFilters[ index ] = {
 			...newActiveFilters[ index ],